moved detailed questions page, and removed unnessesary code
