+++ src/App.tsx
@@ -1,8 +1,5 @@
-//import React, { useState } from 'react';
 import "./styles/App.css";
-//import { Button, Form } from 'react-bootstrap';
-//import { ButtonToPages } from './ButtonToPages';
-import { LocalLink } from "./pages/DetailedQuestions";
+import DetailedQuestions from "./pages/DetailedQuestions";
 import { Home } from "./pages/Home";
 import { SignUp } from "./pages/SignUp";
 import { Login } from "./pages/Login";
@@ -10,33 +7,13 @@ import { Login } from "./pages/Login";
 import { HashRouter, Routes, Route } from "react-router-dom";
 import BasicQuestions from "./pages/BasicQuestions";
 
-//local storage and API Key: key should be entered in by the user and will be stored in local storage (NOT session storage)
-let keyData = "";
-const saveKeyData = "MYKEY";
-const prevKey = localStorage.getItem(saveKeyData); //so it'll look like: MYKEY: <api_key_value here> in the local storage when you inspect
-if (prevKey !== null) {
-  keyData = JSON.parse(prevKey);
-}
-
 function App() {
-  const [key, setKey] = useState<string>(keyData); //for api key input
-  
-  //sets the local storage item to the api key the user inputed
-  function handleSubmit() {
-    localStorage.setItem(saveKeyData, JSON.stringify(key));
-    window.location.reload(); //when making a mistake and changing the key again, I found that I have to reload the whole site before openai refreshes what it has stores for the local storage variable
-  }
-
-  //whenever there's a change it'll store the api key in a local state called key but it won't be set in the local storage until the user clicks the submit button
-  function changeKey(event: React.ChangeEvent<HTMLInputElement>) {
-    setKey(event.target.value);
-  }
   return (
     <HashRouter>
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/starter_helpi" element={<Home />} />
-        <Route path="/DetailedQuestions" element={<LocalLink />} />
+        <Route path="/DetailedQuestions" element={<DetailedQuestions />} />
         <Route path="/BasicQuestions" element={<BasicQuestions />} />
         <Route path="/signup" element={<SignUp />} />
         <Route path="/login" element={<Login />} />
