+++ src/styles/DetailedQuestions.css
